+++ CHANGELOG.md
@@ -9,3 +9,4 @@ All notable changes to **APK Lab** will be documented in this file.
 ### Added
 
 - DeAssemble/Decode APK file using `Apktool`
+- Rebuild APK file (Right click in `apktool.yml` file)